build: added users table
