+++ client/src/components/users/_partials/UsersTable.tsx
@@ -1,0 +1,14 @@
+function UsersTable() {
+	return (
+		<>
+			<section>
+				<h3>Users</h3>
+			</section>
+			<section>
+				<h3>Friends</h3>
+			</section>
+		</>
+	);
+}
+
+export default UsersTable;
